# Made Poly1305 10% faster