Very simple change in the way the poly1305 functions handle 4-bytes to 32-bit integer conversions.
We de-reference the byte while casted as an unsigned integer and that automatically stores or loads the data as it should be. No need for any operations to be done on the data (shifting, ANDing, ORing, etc..)
The reason why I did not optimize the other load and store functions instead of just optimizing Poly1305 was because I used "make speed" and the only code that gets speed benefits is Poly1305, everything else is either the same or regresses.

Results using "make speed"
1490 MB/s --> 1632 MB/s (AMD Ryzen Threadripper 1920X 2x16GB ECC RAM, CPU 4.1GHz, MEM 3GHz)
1334 MB/s --> 1470 MB/s (AMD Ryzen 5 2500U 2x8GB RAM, CPU 2.2GHz Turbo 3.6GHz, MEM 2.4GHz)
1281 MB/s --> _1244 MB/s_ (Intel i7 4500U 2x4GB RAM, CPU 1.7GHz Turbo 3.0 GHz, MEM 2.133GHz)
1461 MB/s --> 1461 MB/s (iMac 5K Late 2015, i5 6600 3.3GHz Turbo 3.9GHz, 4x16GB RAM 1.866GHz)

It seems the change favours AMD's Zen architecture more than Intel's, I will investigate further once I have more hardware on hand, but you can test this yourself by cloning and using "make speed"

Here is code that compares the two ways of converting, and you can clearly see that my version only needs 3 "mov" operations while the older code needs about 30 when you use https://godbolt.org


[examplecode.c.zip](https://github.com/LoupVaillant/Monocypher/files/2994487/examplecode.c.zip)