@Sadoon-AlBader 2019-03-22
> clang -Wcast-align is not happy:
> 
> ```
> examplecode.c:23:13: warning: cast from 'ubyte *' (aka 'unsigned char *') to 'uint *' (aka 'unsigned int *') increases required
>       alignment from 1 to 4 [-Wcast-align]
>   uint b = *(uint*)a;
>             ^~~~~~~~
> ```
> 
> but you can do like
> 
> ```
> static u32 load_u32( const u8 * buf ) {
>     u32 res;
>     memcpy( &res, buf, sizeof( res ) );
>     return res;
> }
> ```
> 
> and compilers are smart enough to optimise that.
> 
> I believe this also needs some extra code to handle big endian platforms
> 
> Also if that second block is a 16 byte load, memcpy( 16 ) should optimise down to movdqu

Sorry, close was by mistake.

I will try your function and see if it gives similar results. Thank you.