@tankf33der 2019-03-22
1. We all see you know tricks and care about CPU ticks, this is good, but you **broke** HPUX (PA-RISC) platform. Expected, right? 
2. Create special form for poly and ignore chacha and EC where load32_le() also used is strange too.
3. Since I see D programming language in your profile much better create and **maintain** DUB bindings for fun and profit.


👎