@Sadoon-AlBader 2019-03-22
>     1. We all see you know tricks and care about CPU ticks, this is good, but you **broke** HPUX (PA-RISC) platform. Expected, right?

Like I said in the previous comment, I will use a static if to select based on endianness during compile-time.

>     2. Create special form for poly and ignore chacha and EC where load32_le() also used is strange too.

Again, like I said in my pull request comment, in my testing I did not find any speedups in other code.

>     3. Since I see D programming language in your profile much better create and **maintain** DUB bindings for fun and profit.

I assure you I will do that (nor for profit though) soon for the Poly1305 component and after that the full MonoCypher library :)
It's part of my project for my independent study course.
Thank you for your concerns regarding endianness and other code but please read the previous comments before repeating what @mikejsavage already excellently pointed out :)