@mliberty1 2019-03-22
Thank you to Sadoon-AlBader for this work.  Unfortunately, I think that your work on optimizing this library misunderstands the goal of Monocypher.  The advantage of Monocypher is that is safe, readable, and completely portable by default.  If you really want speed, use libsodium or any number of other optimized crypto libraries.

The ARM Cortex-M architectures will bus fault if you attempt an unaligned (u32) access.  Now, in this case, the key should be 32-bit aligned, and all accesses appear to be at 4-byte offsets, so remain aligned.  However, this change is potentially dangerous without thorough testing on all target platforms, which is untestable since you don't know!  By keeping the code within the realm of defined C behavior, these operations remain safe.

Personally, I am not interested in Monocypher speed improvements that sacrifice portability or safety.