@Sadoon-AlBader 2019-03-22
Thank you @mliberty1  for your explanation, I now understand better why this optimization is not suitable for Monocypher. Have a nice weekend :)