@LoupVaillant 2019-03-22
Hi, @Sadoon-AlBader,

Don't feel too bad about this PR, I was delighted to even see such an attempt. Yes, this PR is not portable (fails on big endian architectures), and unsafe (your load is actually undefined behaviour). Still, it would seem my version does not generate optimal code, which is a bit of a problem. I'll try and see if there's a way to make it better…

What do you know, I managed to get a 5% increase just by _rolling_ the loading loop:

    diff --git a/src/monocypher.c b/src/monocypher.c
    index 55a0d4c..efa04f3 100644
    --- a/src/monocypher.c
    +++ b/src/monocypher.c
    @@ -388,10 +388,9 @@ void crypto_poly1305_update(crypto_poly1305_ctx *ctx,
         // Process the message block by block
         size_t nb_blocks = message_size >> 4;
         FOR (i, 0, nb_blocks) {
    -        ctx->c[0] = load32_le(message +  0);
    -        ctx->c[1] = load32_le(message +  4);
    -        ctx->c[2] = load32_le(message +  8);
    -        ctx->c[3] = load32_le(message + 12);
    +        FOR (i, 0, 4) {
    +            ctx->c[i] = load32_le(message +  i*4);
    +        }
             poly_block(ctx);
             message += 16;
         }

There's a chance the loop helps with vectorisation. Anyway, thank you for pointing out this low hanging fruit. I wouldn't have seen it without your attempt. I'll push the patch right away.