@mliberty1 2019-03-22
@Sadoon-AlBader,
Optimizing a library like this for a specific platform is a great learning experience!  You may have already learned something about the nasty underbelly of C - undefined behavior.  If you continue to investigate, you will likely find a lot of platform-specific optimizations that cannot be upstreamed.  However, you will also likely find some safe & portable optimizations that can be upstreamed.  I am sure we would all welcome these optimizations!