@Sadoon-AlBader 2019-03-22
> There's a chance the loop helps with vectorisation. Anyway, thank you for pointing out this low hanging fruit. I wouldn't have seen it without your attempt. I'll push the patch right away.

I'm glad I helped in some way :)
Given that this is my first PR I think it could have gone much worse xD
I am actually implementing Poly1305 on hardware (using VHDL so far) and your library helped me a lot in understanding the algorithm and how to optimize it (especially the remainder part), thank you!

> @Sadoon-AlBader,
> Optimizing a library like this for a specific platform is a great learning experience! You may have already learned something about the nasty underbelly of C - undefined behavior. If you continue to investigate, you will likely find a lot of platform-specific optimizations that cannot be upstreamed. However, you will also likely find some safe & portable optimizations that can be upstreamed. I am sure we would all welcome these optimizations!

Thank you, indeed I am learning a lot, and thank you all for the welcoming comments, much appreciated :)