@mliberty1 2019-03-22
@Sadoon-AlBader I am very interested in your VHDL implementation.  I thought about creating an open source Poly1305 and Blake2b FPGA implementation for my [Joulescope project](https://www.joulescope.com/), but decided to leave the authentication/decryption to the microcontroller due to time.  If your work is open source, please let me know where I can follow along.